fix: long word having white spaces when it does not have
Normaly it Would go diagonally the letters, because it allows seeing
different letter in a different perspective.

But if the character is whitespace there is no perspective everything
looks the same, so if you added in a diagonal the long word would have
whitespaces even if it doesn't.

The whitespaces must be added behind the long word